mcp_servers/servo/requirements.txt
mcp
httpx
python-dotenv
zeroconf
uvicorn
